// src/util/types.ts
export interface Token {
    address: string;
    decimals: number;
    name: string;
    symbol: string;
    primaryColor: string;
    id?: string;
    c_id?: string; // coingecko id
    icon?: string;
    displayDecimals: number;
    minAmount?: number;
    maxAmount?: number;
    precision?: number;
    website?: string;
    description?: string;
    verisafe_sticker?: string;
    price_usd?: BigNumber | null;
    price_usd_24h_change?: BigNumber | null;
    listed: boolean;
    isStableCoin: boolean;
}
// ...
export interface TokenBalance {
    balance: BigNumber;
    isUnlocked: boolean;
    token: Token;
}
// ...
export enum ButtonVariant {
    Balance = 'balance',
    Buy = 'buy',
    Error = 'error',
    Primary = 'primary',
    Quaternary = 'quaternary',
    Secondary = 'secondary',
    Sell = 'sell',
    Tertiary = 'tertiary',
    Portis = 'portis',
    Torus = 'torus',
    Fortmatic = 'fortmatic',
}

export enum ButtonIcons {
    Warning = 'warning',
}